==> 2.4_c2017_master_cleaning_soph.Rmd <==
---
title: "Clean Sophomore Spring Survey Master Dataset (for Tamara)"
author: "Juan Ospina"
date: "11/2/2018"
output: 
  html_document:
    toc: TRUE
    toc_float: TRUE
    number_sections: TRUE
---

```{r clean system, message=TRUE, warning=TRUE, include=FALSE}
rm(list = ls()) # clean environment
```

```{r global options, echo=FALSE}
knitr::opts_chunk$set(
  # fig.path = 'figs/', fig.show = 'asis', dpi = 300, 
  # include = FALSE, 
  echo = FALSE, # run the code, show me only the graphs
  warning = FALSE, message = FALSE, cache = FALSE)
```

```{r packages}
library(magrittr)
library(Hmisc)
library(RColorBrewer)
library(reshape2)
# library(MASS) # negative binomial models
library(corrplot)
library(tidyverse)
library(emmeans) # to get adjusted means and to plot the summary statistics of models
library(kableExtra)
library(psych)
library(haven) # package to import SPSS files in R. Useful to import SPSS files and export them as csv

```

```{r}
d0 <- read_rds("../Newest Data/FIntv_2017_Comprehensive_withY1SpGradeData_withSYFUSNoT7_Cleaned_Demogs_Frosh.rds")

# Read SPSS file "FIntv_2017_Soph_Survey_from_Qualtrics_8Nov2018.sav"
d.s <- read_spss("../Newest Data/FIntv_2017_Soph_Survey_from_Qualtrics_8Nov2018.sav")
```

```{r load custom functions}
source("../../../premat_intervention_jo/Research Assistants/scripts/custom_functions.R")
```

# General Notes

## Nomenclature of correlation tables

*   *p* < .001 = ***
*   *p* < .01 = **
*   *p* < .05 = *
*   *p* < 0.1 = † (marginal)
*   *p* < 0.15 = . (trending)

# Review IDs to merge with master dataset

```{r}
# d0 %>% 
#   select(contains("email"), contains("sunet"))

d0 %<>% 
  mutate(
    email = Emailx
  )
```


```{r}
# d.s %>% 
#   names %>% 
#   noquote

# d.s %>% 
#   select(RecipientEmail, email.0, sunetid, freshman, fintv, multiplestudies, physics_wave, physics)

d.s %<>% 
  mutate(
    email = RecipientEmail
  )

count(d.s, fintv)
```

```{r}
# add t4 to the dataset
colnames(d.s) <- paste("t5", colnames(d.s), sep = "")

# head(d.s)

d.s %<>% 
  rename(
    email = t5email
  )

d1 <- d0 %>% 
  left_join(d.s, by = "email")
```

```{r}
d1 %>% 
  count(t5fintv)
```

# Review measures

All partcipants from c2017 were successfully merged!

```{r}
# these are the participants that are not from c2017
# setdiff(
#   d.s %>% select(starts_with("t5friends_kinds_race_friend")), 
#   d1 %>% select(starts_with("t5friends_kinds_race_friend"))
#   )

# all participants were successfully merged!
setdiff(
  d1 %>% select(starts_with("t5friends_kinds_race_friend")),
  d.s %>% select(starts_with("t5friends_kinds_race_friend"))
  )
```

# Measures

## Friendships

Race of participants:

```{r}
d1 %>% 
  count(race)
```

Race labels of their friends:

```{r}
d1 %>% 
  select(
    contains("t5friends_kinds_race_friend")
  ) %>% 
  names %>% 
  noquote

# d.s %>% select(raceid)

attr(d1$t5friends_kinds_race_friend1, "labels")
```

### Friendship Diversity Network
```{r}
d_f <- d1 %>% 
  gather(t5friends_kinds_race_friend1:t5friends_kinds_race_friend7, key = friend, value = friend_race) %>% 
  select(id, race, friend, friend_race) %>%
  mutate(friend = ifelse(friend == "t5friends_kinds_race_friend1", 1, # relabel friend levels
                    ifelse(friend == "t5friends_kinds_race_friend2", 2,
                           ifelse(friend == "t5friends_kinds_race_friend3", 3,
                                  ifelse(friend == "t5friends_kinds_race_friend4", 4,
                                         ifelse(friend == "t5friends_kinds_race_friend5", 5,
                                                ifelse(friend == "t5friends_kinds_race_friend6", 6,
                                                       ifelse(friend == "t5friends_kinds_race_friend7", 7, friend
  )))))))) %>% 
  arrange(id) %>% 
   mutate(
    friend_asian = ifelse(friend_race == 1, 1, 0),
    friend_black = ifelse(friend_race == 2, 1, 0),
    friend_hispanic = ifelse(friend_race == 3, 1, 0),
    friend_native = ifelse(friend_race == 4, 1, 0),
    friend_white = ifelse(friend_race == 5, 1, 0),
    friend_international = ifelse(friend_race == 6, 1, 0),
    friend_multirace = ifelse(friend_race == 7, 1, 0),
    friend_other = ifelse(friend_race == 8, 1, 0)
  )
```

```{r}
attr(d1$t5friends_kinds_race_friend1, "labels")

d_f %>% 
  filter(!is.na(friend_race)) %>% 
  count(friend_race)
  
d_f %>% 
  filter(!is.na(friend_race)) %>% 
  count(race)
```

```{r}
d_f %<>%
  mutate(friend_race_f = 
           ifelse(
             friend_race == 1, "asian",
             ifelse(
               friend_race == 2, "black",
               ifelse(
                 friend_race == 3, "latino",
                 ifelse(
                   friend_race == 4, "native.pi",
                   ifelse(
                     friend_race == 5, "white",
                     ifelse(
                       friend_race == 6, "international",
                       ifelse(
                         friend_race == 7, "multiracial", "other"
                       )
                     )
                   )
                 )
               )
             )
           )
  )

d_f %>%
  count(friend_race_f, friend_race)

d_f %<>% 
  mutate(
    friend_race_same = ifelse(race == friend_race_f, 1, 0),
    friend_race_cross = ifelse(race != friend_race_f, 1, 0)
    )

# d_f %>% 
#   count(race, friend_race_f, friend_race_same)


d_f_race <-
  d_f %>% 
  filter(!is.na(friend_race)) %>% # For the count
  group_by(id) %>% 
  summarise(friend_total = n(),
            friend_asian_count =  sum(friend_asian),
            friend_black_count =  sum(friend_black),
            friend_hispanic_count =  sum(friend_hispanic),
            friend_native_count =  sum(friend_native),
            friend_white_count =  sum(friend_white),
            friend_international_count =  sum(friend_international),
            friend_multirace_count =  sum(friend_multirace),
            friend_other_count =  sum(friend_other),
            friend_race_same_count = sum(friend_race_same),
            friend_race_cross_count = sum(friend_race_cross)
            ) %>% 
  ungroup() %>% 
  mutate(
    friend_asian_pct = friend_asian_count/friend_total*100,
    friend_black_pct = friend_black_count/friend_total*100,
    friend_hispanic_pct = friend_hispanic_count/friend_total*100,
    friend_native_pct = friend_native_count/friend_total*100,
    friend_white_pct = friend_white_count/friend_total*100,
    friend_international_pct = friend_international_count/friend_total*100,
    friend_multirace_pct = friend_multirace_count/friend_total*100,
    friend_other_pct = friend_other_count/friend_total*100,
    friend_race_same_pct = friend_race_same_count/friend_total*100,
    friend_race_cross_pct = friend_race_cross_count/friend_total*100)

d_f_race %>% 
  rowwise() %>% 
  mutate(total_pct = sum(friend_asian_pct, friend_black_pct, friend_hispanic_pct, friend_native_pct, friend_white_pct, friend_international_pct, friend_multirace_pct, friend_other_pct)) %>%
  ungroup() %>% 
  count(total_pct)

# add t4 to the dataset
colnames(d_f_race) <- paste("t5", colnames(d_f_race), sep = "")

d_f_race %<>% 
  rename(
    id = t5id
  )
```

```{r}
d1 <- d_f_race %>% 
  right_join(d1, by = "id")


d1 %>% count(t4friend_total)

d1 %>% count(t4friend_asian_count)
d1 %>% count(t4friend_black_count)
d1 %>% count(t4friend_asian_pct)
d1 %>% count(t4friend_black_pct)
d1 %>% count(t4friend_hispanic_pct)
d1 %>% count(t4friend_native_pct)
d1 %>% count(t4friend_white_pct)
d1 %>% count(t4friend_international_pct)
d1 %>% count(t4friend_multirace_pct)
d1 %>% count(t4friend_other_pct)
d1 %>% count(t4friend_race_same_pct)
d1 %>% count(t4friend_race_cross_pct)

# d0 %>% select(race, starts_with("friend"))

d_f_race %>% 
  select(-id) %>% 
  names %>% 
  noquote
```

# Started Sophomore Survey

```{r}
check_vars_by_keywords(d1, c("start"))

d2 <- d1 %>% filter(!is.na(t5StartDate))
d1 %<>% 
  mutate(
    startedSOPH = ifelse(!is.na(t5StartDate), 1, 0)
  )

d1 %>% count(startedSOPH)
```


# Export dataset
```{r}
d1 %>% write_rds("../Newest Data/FIntv_2017_Comprehensive_withY1SpGradeData_withSYFUSNoT7_Cleaned_Demogs_Frosh_Soph.rds")
d1 %>% write_csv("../Newest Data/FIntv_2017_Comprehensive_withY1SpGradeData_withSYFUSNoT7_Cleaned_Demogs_Frosh_Soph.csv")
```
