==> 2.5_c2017_master_cleaning_syfus_tamara.Rmd <==
---
title: "Clean Senior Spring Survey Master Dataset (for Tamara)"
author: "Juan Ospina"
date: "11/2/2018"
output: 
  html_document:
    toc: TRUE
    toc_float: TRUE
    number_sections: TRUE
---

```{r clean system, message=TRUE, warning=TRUE, include=FALSE}
rm(list = ls()) # clean environment
```

```{r global options, echo=FALSE}
knitr::opts_chunk$set(
  # fig.path = 'figs/', fig.show = 'asis', dpi = 300, 
  # include = FALSE, 
  echo = FALSE, # run the code, show me only the graphs
  warning = FALSE, message = FALSE, cache = FALSE)
```

```{r packages}
library(magrittr)
library(Hmisc)
library(RColorBrewer)
library(reshape2)
# library(MASS) # negative binomial models
library(corrplot)
library(tidyverse)
library(emmeans) # to get adjusted means and to plot the summary statistics of models
library(kableExtra)
library(psych)
```

```{r}
d0 <- read_rds("../Newest Data/FIntv_2017_Comprehensive_withY1SpGradeData_withSYFUSNoT7_Cleaned_Demogs_Frosh_Soph.rds")
```

```{r load custom functions}
source("../../../premat_intervention_jo/Research Assistants/scripts/custom_functions.R")
```

# General Notes

## Nomenclature of correlation tables

*   *p* < .001 = ***
*   *p* < .01 = **
*   *p* < .05 = *
*   *p* < 0.1 = † (marginal)
*   *p* < 0.15 = . (trending)

# Measures

## Independence Valued at College

Review variables with keyword "qualities_ind":

```{r}
check_vars_by_keywords(d0, c("qualities_ind"))
```

Items and scales:

*     *qualities_indgoals*:

```{r}
attr(d0$qualities_indgoals, "label")
attr(d0$qualities_indgoals, "labels")
```

*     *qualities_indownpath*:

```{r}
attr(d0$qualities_indownpath, "label")
attr(d0$qualities_indownpath, "labels")
```

*     *qualities_indselfmotivated*:

```{r}
attr(d0$qualities_indselfmotivated, "label")
attr(d0$qualities_indselfmotivated, "labels")
```

**Correlations**:

Since *qualities_indownpath* is weakly correlated, the composite *qualitiesind_comp2* includes only the average between *qualities_indgoals* and *qualities_indselfmotivated*. Scale reliability with the three measures is also low and factor analyses too (see below).

```{r}
vars_indep <- quos(qualitiesind_comp2, qualities_indgoals, qualities_indownpath, qualities_indselfmotivated)
corr_table_p(d0, vars_indep)
```

**Scale Reliability**:

```{r}
vars_indep <- quos(qualities_indgoals, qualities_indownpath, qualities_indselfmotivated)
alphatize_2(d0, vars_indep)
```

**Correlation Plot**:

```{r}
corr_plot(d0, vars_indep)
```

**Factor Analysis**:

```{r}
factor_analysis(d0, vars_indep, 1)
```


## Interdependence Valued at College

Review variables with keyword "qualities_inter":

```{r}
check_vars_by_keywords(d0, c("qualities_inter"))
```

Items and scales:

*     *qualities_intersharedvalues*:

```{r}
attr(d0$qualities_intersharedvalues, "label")
attr(d0$qualities_intersharedvalues, "labels")
```

*     *qualities_interconnected*:

```{r}
attr(d0$qualities_interconnected, "label")
attr(d0$qualities_interconnected, "labels")
```

*     *qualities_intergenrel*:

```{r}
attr(d0$qualities_intergenrel, "label")
attr(d0$qualities_intergenrel, "labels")
```

*     *qualities_intergiveback*:

```{r}
attr(d0$qualities_intergiveback, "label")
attr(d0$qualities_intergiveback, "labels")
```

*     *qualities_interothersmotivated*:

```{r}
attr(d0$qualities_interothersmotivated, "label")
attr(d0$qualities_interothersmotivated, "labels")
```

**Correlations**:

Since *qualities_interothersmotivated* is not correlated, the composite *qualitiesinter_comp4* includes only the average between *qualities_intersharedvalues*, *qualities_interconnected*, *qualities_intergenrel*, and *qualities_intergiveback*. Scale reliability with the four measures is higher and factor analyses too (see below).

```{r}
vars_interdep <- 
  quos(
    qualities_intersharedvalues, qualities_interconnected, qualities_intergenrel,                   
    qualities_intergiveback, qualities_interothersmotivated
)
corr_table_p(d0, vars_interdep)
```

**Correlation Plot**:

```{r}
corr_plot(d0, vars_interdep)
```

**Scale Reliability**:

```{r}
alphatize_2(d0, vars_interdep)
```

**Factor Analysis**:

```{r}
factor_analysis(d0, vars_interdep, 1)
```

**Scale Reliability** (without *qualities_interothersmotivated*):

```{r}
vars_interdep <- 
  quos(
    qualities_intersharedvalues, qualities_interconnected, qualities_intergenrel,                   
    qualities_intergiveback
)
alphatize_2(d0, vars_interdep)
```

**Factor Analysis** (without *qualities_interothersmotivated*):

```{r}
factor_analysis(d0, vars_interdep, 1)
```

## Independent purpose for going to college

Review variables with keyword "t4purpose":

```{r}
vars_interdep <- 
  quos(
    t4interdep1, t4interdep2, t4interdep3, t4interdep4
)
check_vars_by_keywords(d0, c("t4purpose"))
```

Items and scales:

*     *t4purpose_earnmoney*:

```{r}
attr(d0$t4purpose_earnmoney, "label")
attr(d0$t4purpose_earnmoney, "labels")
```

*     *t4purpose_proofcase*:

```{r}
attr(d0$t4purpose_proofcase, "label")
attr(d0$t4purpose_proofcase, "labels")
```

*     *t4purpose_positiveimpact*:

```{r}
attr(d0$t4purpose_positiveimpact, "label")
attr(d0$t4purpose_positiveimpact, "labels")
```

*     *t4purpose_edcitizen*:

```{r}
attr(d0$t4purpose_edcitizen, "label")
attr(d0$t4purpose_edcitizen, "labels")
```

*     *t4purpose_expandknowledge*:

```{r}
attr(d0$t4purpose_expandknowledge, "label")
attr(d0$t4purpose_expandknowledge, "labels")
```

*     *t4purpose_indepthinker*:

```{r}
attr(d0$t4purpose_indepthinker, "label")
attr(d0$t4purpose_indepthinker, "labels")
```

*     *t4purpose_learnabtinterests*:

```{r}
attr(d0$t4purpose_learnabtinterests, "label")
attr(d0$t4purpose_learnabtinterests, "labels")
```

*     *t4purposeindepcomp7*:

```{r}
attr(d0$t4purposeindepcomp7, "label")
```

**Correlations**:

```{r}
vars_indep <- 
  quos(
    t4purposeindepcomp7, t4purpose_earnmoney, t4purpose_proofcase, t4purpose_positiveimpact, 
    t4purpose_edcitizen, t4purpose_expandknowledge, t4purpose_indepthinker, t4purpose_learnabtinterests
)

corr_table_p(d0, vars_indep)

vars_indep <- 
  quos(
    t4purpose_earnmoney, t4purpose_proofcase, t4purpose_positiveimpact, 
    t4purpose_edcitizen, t4purpose_expandknowledge, t4purpose_indepthinker, t4purpose_learnabtinterests
)
```

**Scale Reliability**:

```{r}
alphatize_2(d0, vars_indep)
```

**Factor Analysis**:

```{r}
factor_analysis(d0, vars_indep, 1)
```

**Scale Reliability**: Dropping *t4purpose_earnmoney* and *t4purpose_proofcase*.

```{r}
vars_indep <- 
  quos(
    t4purpose_positiveimpact, t4purpose_edcitizen, t4purpose_expandknowledge, 
    t4purpose_indepthinker, t4purpose_learnabtinterests
)

alphatize_2(d0, vars_indep)
```

**Factor Analysis**: Dropping *t4purpose_earnmoney* and *t4purpose_proofcase*.

```{r}
factor_analysis(d0, vars_indep, 1)
```


```{r}
d0 <- cv_compute(d0, cv_name = t4purposeindepcomp5, cv_vector = vars_indep)
```

## Friendships

Race of participants:

```{r}
d0 %>% 
  count(race)
```

Race labels of their friends:

```{r}
d0 %>% 
  select(
    starts_with("friend_race")
    # ends_with("race")
  )

attr(d0$friend_race_friend1, "labels")
```

### Friendship Diversity Network
```{r}
# % of black students should be a multiple of 7 or bigger than 14%
# subset data to counts of friends (7 friends, 6 friends)

d_f <- d0 %>% 
  gather(friend_race_friend1:friend_race_friend7, key = friend, value = friend_race) %>% 
  select(id, race, friend, friend_race) %>%
  mutate(friend = ifelse(friend == "friend_race_friend1", 1, # relabel friend levels
                    ifelse(friend == "friend_race_friend2", 2,
                           ifelse(friend == "friend_race_friend3", 3,
                                  ifelse(friend == "friend_race_friend4", 4,
                                         ifelse(friend == "friend_race_friend5", 5,
                                                ifelse(friend == "friend_race_friend6", 6,
                                                       ifelse(friend == "friend_race_friend7", 7, friend
  )))))))) %>% 
  arrange(id) %>% 
  mutate(
    friend_asian = ifelse(friend_race == 1, 1, 0),
    friend_black = ifelse(friend_race == 2, 1, 0),
    friend_hispanic = ifelse(friend_race == 3, 1, 0),
    friend_native = ifelse(friend_race == 4, 1, 0),
    friend_mideast = ifelse(friend_race == 5, 1, 0),
    friend_multirace = ifelse(friend_race == 6, 1, 0),
    friend_pi = ifelse(friend_race == 7, 1, 0),
    friend_white = ifelse(friend_race == 8, 1, 0),
    friend_other = ifelse(friend_race == 9, 1, 0)
  )
```

```{r}
d_f %>% 
  filter(!is.na(friend_race)) %>% 
  count(friend_race)
  
d_f %>% 
  filter(!is.na(friend_race)) %>% 
  count(race)
```

```{r}
d_f %<>%
  mutate(friend_race_f = 
           ifelse(
             friend_race == 1, "asian",
             ifelse(
               friend_race == 2, "black",
               ifelse(
                 friend_race == 3, "latino",
                 ifelse(
                   friend_race == 4, "native.pi",
                   ifelse(
                     friend_race == 5, "middle east",
                     ifelse(
                       friend_race == 6, "multiracial",
                       ifelse(
                         friend_race == 7, "native.pi", 
                         ifelse(
                           friend_race == 8, "white", "other"
                         )
                       )
                     )
                   )
                 )
               )
             )
           )
  )

d_f %>%
  count(friend_race_f, friend_race)

d_f %<>% 
  mutate(
    friend_race_same = ifelse(race == friend_race_f, 1, 0),
    friend_race_cross = ifelse(race != friend_race_f, 1, 0)
    )

# d_f %>%
#   count(race, friend_race_f, friend_race_same)

# friend_asian = ifelse(friend_race == 1, 1, 0),
# friend_black = ifelse(friend_race == 2, 1, 0),
# friend_hispanic = ifelse(friend_race == 3, 1, 0),
# friend_native = ifelse(friend_race == 4, 1, 0),
# friend_mideast = ifelse(friend_race == 5, 1, 0),
# friend_multirace = ifelse(friend_race == 6, 1, 0),
# friend_pi = ifelse(friend_race == 7, 1, 0),
# friend_white = ifelse(friend_race == 8, 1, 0),
# friend_other = ifelse(friend_race == 9, 1, 0)

d_f_race <-
  d_f %>% 
  filter(!is.na(friend_race)) %>% # For the count
  group_by(id) %>% 
  summarise(friend_total = n(),
            friend_asian_count =  sum(friend_asian),
            friend_black_count =  sum(friend_black),
            friend_hispanic_count =  sum(friend_hispanic),
            friend_native_count =  sum(friend_native),
            friend_mideast_count =  sum(friend_mideast),
            friend_multirace_count =  sum(friend_multirace),
            friend_pi_count =  sum(friend_pi),
            friend_white_count =  sum(friend_white),
            friend_other_count =  sum(friend_other),
            friend_race_same_count = sum(friend_race_same),
            friend_race_cross_count = sum(friend_race_cross)
            ) %>% 
  ungroup() %>% 
  mutate(
    friend_asian_pct = friend_asian_count/friend_total*100,
    friend_black_pct = friend_black_count/friend_total*100,
    friend_hispanic_pct = friend_hispanic_count/friend_total*100,
    friend_native_pct = friend_native_count/friend_total*100,
    friend_mideast_pct = friend_mideast_count/friend_total*100,
    friend_multirace_pct = friend_multirace_count/friend_total*100,
    friend_pi_pct = friend_pi_count/friend_total*100,
    friend_white_pct = friend_white_count/friend_total*100,
    friend_other_pct = friend_other_count/friend_total*100,
    friend_race_same_pct = friend_race_same_count/friend_total*100,
    friend_race_cross_pct = friend_race_cross_count/friend_total*100)

d_f_race %>% 
  rowwise() %>% 
  mutate(total_pct = sum(friend_asian_pct, friend_black_pct, friend_hispanic_pct, friend_native_pct, friend_mideast_pct, friend_multirace_pct, friend_pi_pct, friend_white_pct, friend_other_pct)) %>%
  ungroup() %>% 
  count(total_pct)

# add t4 to the dataset
colnames(d_f_race) <- paste("t7", colnames(d_f_race), sep = "")

d_f_race %<>% 
  rename(
    id = t7id
  )
```

```{r}
d0 <- d_f_race %>% 
  right_join(d0, by = "id")


d0 %>% count(t7friend_total)

d0 %>% count(t7friend_asian_count)
d0 %>% count(t7friend_black_count)
d0 %>% count(t7friend_asian_pct)
d0 %>% count(t7friend_black_pct)
d0 %>% count(t7friend_hispanic_pct)
d0 %>% count(t7friend_native_pct)
d0 %>% count(t7friend_white_pct)
d0 %>% count(t7friend_mideast_pct)
d0 %>% count(t7friend_multirace_pct)
d0 %>% count(t7friend_other_pct)
d0 %>% count(t7friend_race_same_pct)
d0 %>% count(t7friend_race_cross_pct)

# d0 %>% select(race, starts_with("friend"))

d_f_race %>% 
  select(-id) %>% 
  names %>% 
  noquote
```

# Export dataset
```{r}
d0 %>% write_rds("../Newest Data/FIntv_2017_Comprehensive_withY1SpGradeData_withSYFUSNoT7_Cleaned_Demogs_Frosh_Soph_Senior.rds")
d0 %>% write_csv("../Newest Data/FIntv_2017_Comprehensive_withY1SpGradeData_withSYFUSNoT7_Cleaned_Demogs_Frosh_Soph_Senior.csv")
```
